+++ seismiqb/__init__.py
@@ -9,8 +9,8 @@ from .field import Field, SyntheticField
 from .geometry import SeismicGeometry, BloscFile, array_to_sgy
 from .labels import Horizon, UnstructuredHorizon, Fault, GeoBody
 from .metrics import HorizonMetrics, GeometryMetrics, FaultsMetrics, FaciesMetrics
-from .samplers import GeometrySampler, HorizonSampler, FaultSampler, ConstantSampler, \
-                      SeismicSampler, BaseGrid, RegularGrid, ExtensionGrid, LocationsPotentialContainer
+from .samplers import GeometrySampler, HorizonSampler, FaultSampler, ConstantSampler, SeismicSampler
+from .grids import  BaseGrid, RegularGrid, ExtensionGrid, LocationsPotentialContainer
 
 # Utilities and helpers
 from .plotters import MatplotlibPlotter, plot_image, plot_loss
