+++ seismiqb/utils/__init__.py
@@ -1,5 +1,7 @@
 """ Utility functions and classes. """
 # pylint: disable=wildcard-import
+from .accumulator2d import *
+from .accumulator3d import *
 from .cache import *
 from .charisma import *
 from .classes import *
